2.04.2021 3.1.2 Финал.. на проверку!

+++ src/main/java/com/oschepich/spring_boot/bootstrap/controler/AdminController.java
@@ -22,34 +22,42 @@ public class AdminController {
     private final UserService userService;
     private final UserRepository userRepository;
 
+
     public AdminController(UserService userService, UserRepository userRepository) {
         this.userService = userService;
         this.userRepository = userRepository;
+
     }
 
     //вывожу всех пользователей
     @GetMapping
     public ModelAndView getAllUsers(ModelAndView modelAndView, Principal principal) {
-        modelAndView.addObject( "tecUser", this.userRepository.findUserByEmail(principal.getName()));
+        modelAndView.addObject("tecUser", this.userRepository.findUserByEmail(principal.getName()));
         modelAndView.addObject("allUser", userService.getAllUser());
         modelAndView.addObject("user", new User());
         modelAndView.setViewName("admin");
         return modelAndView;
     }
+
     //создаю нового пользователя
     @GetMapping("/new")
     public String addNewUser(Model model) {
         model.addAttribute("user", new User());
         return "admin";
     }
+
     //создаю нового пользователя
     @PostMapping("/add")
     public String addUser(@ModelAttribute("user") User user,
                           @RequestParam(value = "roleSet", required = false) String[] roleSet) {
-        user.setRoles(getAddRole(roleSet));
-        userService.saveUser(user);
-        return "redirect:/admin";
+        if (userRepository.findUserByEmail(user.getEmail()) == null) {
+            user.setRoles(getAddRole(roleSet));
+            userService.saveUser(user);
+            return "redirect:/admin";
+        }
+        return "access";
     }
+
     //редактирую пользователя
     @GetMapping("/edit/{id}")
     public ModelAndView getUserToEdit(@PathVariable Long id) {
@@ -59,27 +67,27 @@ public class AdminController {
         modelAndView.addObject("user", user);
         return modelAndView;
     }
+
     //редактирую пользователя
     @PostMapping("/edit/{id}")
     public String editUser(@ModelAttribute("user") User user,
                            @RequestParam(value = "roleSet", required = false) String[] roleSet) {
-            user.setRoles(getAddRole(roleSet));
-            userService.saveUser(user);
+        user.setRoles(getAddRole(roleSet));
+        userService.saveUser(user);
         return "redirect:/admin";
     }
+
     //удаляю пользователя
     @PostMapping("/delete/{id}")
     public String deleteUser(@PathVariable Long id, Principal principal) {
+        if (userRepository.findUserByEmail(principal.getName()).getId().equals(id)) {
+            return "access";
+        }
         userService.deleteUser(id);
         return "redirect:/admin";
     }
 
     private Set<Role> getAddRole(String[] roles) {
-//        if (roles == null){
-//            Set<Role> roleSet = new HashSet<>();
-//            roleSet.add(userService.getRoleByName("ROLE_USER"));
-//            return roleSet;
-//        }
         Set<Role> roleSet = new HashSet<>();
         for (String role : roles) {
             roleSet.add(userService.getRole(role));
